pyproject.toml
[project]
name = "layer-ai"
version = "0.1.0"
description = "Single-image layer extraction and motion-ready asset packaging"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "jsonschema>=4.23,<5.0",
  "pillow>=11.2,<12.0",
  "pydantic>=2.11,<3.0",
  "pydantic-settings>=2.8,<3.0",
  "python-multipart>=0.0.20,<0.1.0",
  "uvicorn>=0.34,<0.35",
]

[dependency-groups]
dev = [
  "httpx>=0.28,<0.29",
  "pytest>=8.3,<9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages/layer_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["packages"]

